feat(frames): move frame actions under the canvas
